:sparkles: :: cookie setup

+++ src/main/java/com/msg/after_school/domain/auth/data/dto/TokenDto.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class TokenDto {
     private String accessToken;
     private String refreshToken;
-    private Long accessExp;
-    private Long refreshExp;
+    private Integer accessExp;
+    private Integer refreshExp;
 }
